libexec/cksum: add --no-check-binlog-format flag to pt-table-checksum runs
Ideally, this flag would not be needed - currently it is only being
triggered due to replica DB's not being configured to write statement
based replication logs for their own data, which, in a 1-tier
replication scheme is not relevant to the checks performed when
this flag is not off.

Proper fix would be ensuring replicas use statement based replication for
their replication logs.